offer/trees/issymmetric_test.go
package trees

import (
	"algorithm-go/structure/binarytree"
	"testing"
)

func Test_isSymmetric(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		// TODO: Add test cases.
		{
			"case1",
			args{
				root: binarytree.FromSlice([]int{1, 2, 3}).Root,
			},
			false,
		},
		{
			"case2",
			args{
				root: binarytree.FromSlice([]int{1, 2, 2, 3, 4, 4, 3}).Root,
			},
			true,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := isSymmetric(tt.args.root); got != tt.want {
				t.Errorf("isSymmetric() = %v, want %v", got, tt.want)
			}
		})
	}
}
